requirements.txt
python-dotenv
requests
beautifulsoup4
sqlalchemy
psycopg2-binary
aiohttp
APScheduler
